+++ tests/test-worker/migrate.ts
@@ -10,10 +10,10 @@ worker.logJobs = true;
 export const migrate = async () => {
   await worker.migrateDatabase();
   await worker.register({
-    username: "ricsam",
+    username: "adm1n",
     email: "oss@ricsam.dev",
-    name: "Ricsam",
-    password: "password",
+    name: "Admin",
+    password: "s3cr3t",
     admin: true,
   });
 };