+++ packages/hub/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @enschedule/hub
 
+## 1.0.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @enschedule/worker-api@1.1.2
+  - @enschedule/pg-driver@1.1.2
+  - @enschedule/worker@1.1.2
+  - @enschedule/dashboard@1.1.2
+  - @enschedule/types@1.1.1
+
 ## 1.0.2
 
 ### Patch Changes