+++ packages/types/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @enschedule/types
 
+## 1.1.1
+
+### Patch Changes
+
+- use nafs for storing and reading logs
+
 ## 1.1.0
 
 ### Minor Changes